+++ server/src/rss/rss.module.ts
@@ -3,12 +3,20 @@ import { Module } from '@nestjs/common';
 import { RssController } from './rss.controller';
 import { RssService } from './rss.service';
 import { RssInformation } from './rss.entity';
-import { RssRejectRepository, RssRepository } from './rss.repository';
-import { BlogRepository } from '../blog/blog.repository';
+import {
+  RssRejectRepository,
+  RssRepository,
+  RssAcceptRepository,
+} from './rss.repository';
 
 @Module({
   imports: [TypeOrmModule.forFeature([RssInformation])],
   controllers: [RssController],
-  providers: [RssService, RssRepository, BlogRepository, RssRejectRepository],
+  providers: [
+    RssService,
+    RssRepository,
+    RssAcceptRepository,
+    RssRejectRepository,
+  ],
 })
 export class RssModule {}
