+++ server/src/rss/rss.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { DataSource, Repository } from 'typeorm';
-import { Rss, RssReject } from './rss.entity';
+import { Rss, RssAccept, RssReject } from './rss.entity';
 import { RssRegisterDto } from './dto/rss-register.dto';
 
 @Injectable()
@@ -25,5 +25,12 @@ export class RssRepository extends Repository<Rss> {
 export class RssRejectRepository extends Repository<RssReject> {
   constructor(private dataSource: DataSource) {
     super(RssReject, dataSource.createEntityManager());
+  }
+}
+
+@Injectable()
+export class RssAcceptRepository extends Repository<RssAccept> {
+  constructor(private readonly dataSource: DataSource) {
+    super(RssAccept, dataSource.createEntityManager());
   }
 }
